ending dance

+++ include/PackageData.h
@@ -1,3 +1,5 @@
+// 欧阳承风
+
 #pragma once
 
 #include <stddef.h>
@@ -7,17 +9,18 @@ typedef struct PackageData{
 	// 1xx - xxxx - xxx
 	char name[31];
 	// 1 - 10 个字符
-	double distance;
-	// (0, 10000] km
-	double volume;
+	int distance;
+	// * 100 后 | (0, 10000] km
+	int volume;
 	/*
+		* 100 后 |
 		(0, 120000] cm ^ 3
 		(0, 36000] cm ^ 3 为小型货物
 		(36000, 60000] cm ^ 3 为中型货物
 		(60000, 120000] cm ^ 3 为大型货物
 	*/
-	double weight;
-	// (0, 200] kg ，大于 100 kg 算超重
+	int weight;
+	// * 100 后 | (0, 200] kg ，大于 100 kg 算超重
 	char PlaceFrom[61], PlaceTo[61];
 	// 1 - 20 个字符
 	int status;
@@ -31,4 +34,4 @@ typedef struct PackageData{
 	// 8 个字符，由英文大小写字符和数字组成
 }PackageData;
 
-PackageData CreatePackage(long long, char*, double, double, double, char*, char*, int);
+PackageData CreatePackage(long long, char*, int, int, int, char*, char*, int);
